+++ examples/storybook/stories/Experiments/ImagePluginQualityPreload.tsx
@@ -11,7 +11,6 @@ const ImagePluginStory = () => {
   const [viewerKey, setViewerKey] = useState(1);
 
   const experiments = parseExperiments({
-    'specs.ricos.useQualityPreload': 'true',
     'specs.ricos.useSrcSet': 'true',
   });
   // eslint-disable-next-line no-console