+++ packages/plugin-gallery/web/src/createGalleryPlugin.ts
@@ -1,5 +1,5 @@
 import createToolbar from './toolbar/createToolbar';
-import { createBasePlugin, createBaseMediaPlugin } from 'wix-rich-content-plugin-commons';
+import { createBasePlugin } from 'wix-rich-content-plugin-commons';
 import { Component, DEFAULTS } from './gallery-component';
 import { GALLERY_TYPE, GalleryPluginEditorConfig } from './types';
 import { CreatePluginFunction } from 'wix-rich-content-common';
@@ -45,8 +45,9 @@ const createGalleryPlugin: CreatePluginFunction<GalleryPluginEditorConfig> = con
           disableExpand: settings.disableExpand,
         }
       : pluginData;
+
   return createBasePlugin({
-    component: createBaseMediaPlugin(Component),
+    component: Component,
     settings,
     theme,
     t,