+++ packages/plugin-image/web/src/createImagePlugin.ts
@@ -3,7 +3,6 @@ import {
   createBasePlugin,
   PLUGIN_DECORATION_PROPS,
   PLUGIN_DECORATIONS,
-  createBaseMediaPlugin,
 } from 'wix-rich-content-plugin-commons';
 import { Component, DEFAULTS } from './image-component';
 import { IMAGE_TYPE, IMAGE_TYPE_LEGACY, ImagePluginEditorConfig } from './types';
@@ -35,8 +34,9 @@ const createImagePlugin: CreatePluginFunction<ImagePluginEditorConfig> = config 
     experiments,
     ...rest
   } = config;
+
   return createBasePlugin({
-    component: createBaseMediaPlugin(Component),
+    component: Component,
     type: IMAGE_TYPE,
     legacyType: IMAGE_TYPE_LEGACY,
     pluginDecorationProps: (props, componentData) => {