+++ packages/plugin-video/web/src/types.ts
@@ -12,7 +12,7 @@ export const VIDEO_TYPE = 'wix-draft-plugin-video';
 export const VIDEO_TYPE_LEGACY = 'VIDEO-EMBED';
 
 export const videoButtonsTypes = { video: 'video', soundCloud: 'soundCloud', youTube: 'youTube' };
-export const mediaTypes = [videoButtonsTypes.soundCloud, videoButtonsTypes.youTube];
+export const meidaTypes = [videoButtonsTypes.soundCloud, videoButtonsTypes.youTube];
 export interface VideoPluginEditorConfig extends EditorPluginConfig {
   [key: string]: any; // eslint-disable-line @typescript-eslint/no-explicit-any
 }