+++ CHANGELOG.md
@@ -21,6 +21,8 @@
 
 - `editor-common`
   - [#2663](https://github.com/wix/ricos/pull/2663) moved SearchIcon/ErrorIcon into ui-components folder
+- `image`
+  - []() Merge `useQualityPreload` Experiment
 
 </details>
 <hr/>