+++ CHANGELOG.md
@@ -21,6 +21,8 @@
 
 - `ricos-editor`
   - [#2684](https://github.com/wix-incubator/rich-content/pull/2684) BM's top toolbar overrides Ricos Modal
+- `editor-common`
+  - [#2685](https://github.com/wix-incubator/rich-content/pull/2685) color utils for editor commands
 
 ### :nail_care